backend/requirements.txt
fastapi
python-multipart
uvicorn
docker
psutil
httpx
sqlmodel
passlib
python-jose